backend/medicine-verification-ussd/requirements.txt
fastapi
python-dotenv
supabase
uvicorn
